add function to messageHelpers and add a couple tests
