+++ apps/test/unit/pythonlab/pythonHelpers/messageHelpersTest.ts
@@ -1,0 +1,19 @@
+import {parseMessageToNeighborhoodSignal} from '@cdo/apps/pythonlab/pythonHelpers/messageHelpers';
+
+describe('parseMessageToNeighborhoodSignal', function () {
+  it('can successfully parse a message string with no detail', async function () {
+    expect(parseMessageToNeighborhoodSignal('[PAINTER] MOVE')).toEqual({
+      value: 'MOVE',
+    });
+  });
+  it('can successfully parse a message string with detail', async function () {
+    expect(
+      parseMessageToNeighborhoodSignal('[PAINTER] PAINT {"color": "Blue"}')
+    ).toEqual({
+      value: 'PAINT',
+      detail: {
+        color: 'Blue',
+      },
+    });
+  });
+});
